update divide base case
divide base case when one point or two point or three point
move base case method to VoronoiBaseCase.java

+++ source/src/main/java/dev/dockyu/voronoidiagram/datastruct/GeneratorPoint.java
@@ -1,14 +1,14 @@
 package dev.dockyu.voronoidiagram.datastruct;
 
 public class GeneratorPoint {
-    int x; // x座標
-    int y; // y座標
+    float x; // x座標
+    float y; // y座標
 
-    public GeneratorPoint(int clickedX, int clickedY) {
+    public GeneratorPoint(float clickedX, float clickedY) {
         this.x = clickedX;
         this.y = clickedY;
     }
 
-    public int getX() { return this.x;}
-    public int getY() { return this.y;}
+    public float getX() { return this.x;}
+    public float getY() { return this.y;}
 }
